docs: fixing syntax (#163)
line 151: withdraw - withdrawal
line 162: build -> built
line 162: dAPP -> dApps
line 166: unkonwn -> unknown - on Godwoken Bridge page -> on the Godwoken Bridge page -  this is because that your account is not registered on Godwoken network - this may be because your account is not registered on the Godwoken network. This can be fixed by depositing some CKB to the Godwoken network, then the issue should be resolved.